rename lint, docs, improve diagnostics
