--- tests/ui/implied_bounds_in_impl.stderr
@@ -1,46 +1,0 @@
-error: this bound is implied by another bound and can be removed
-  --> $DIR/implied_bounds_in_impl.rs:12:36
-   |
-LL | fn deref_derefmut<T>(x: T) -> impl Deref<Target = T> + DerefMut<Target = T> {
-   |                                    ^^^^^^^^^^^^^^^^^
-   |
-   = note: `-D clippy::implied-bounds-in-impl` implied by `-D warnings`
-
-error: this bound is implied by another bound and can be removed
-  --> $DIR/implied_bounds_in_impl.rs:27:34
-   |
-LL | fn generics_implied<T>() -> impl GenericTrait<T> + GenericSubtrait<(), T, ()>
-   |                                  ^^^^^^^^^^^^^^^
-
-error: this bound is implied by another bound and can be removed
-  --> $DIR/implied_bounds_in_impl.rs:33:40
-   |
-LL | fn generics_implied_multi<V>() -> impl GenericTrait<i32> + GenericTrait2<V> + GenericSubtrait<(), i32, V> {}
-   |                                        ^^^^^^^^^^^^^^^^^
-
-error: this bound is implied by another bound and can be removed
-  --> $DIR/implied_bounds_in_impl.rs:33:60
-   |
-LL | fn generics_implied_multi<V>() -> impl GenericTrait<i32> + GenericTrait2<V> + GenericSubtrait<(), i32, V> {}
-   |                                                            ^^^^^^^^^^^^^^^^
-
-error: this bound is implied by another bound and can be removed
-  --> $DIR/implied_bounds_in_impl.rs:35:44
-   |
-LL | fn generics_implied_multi2<T, V>() -> impl GenericTrait<T> + GenericTrait2<V> + GenericSubtrait<(), T, V>
-   |                                            ^^^^^^^^^^^^^^^
-
-error: this bound is implied by another bound and can be removed
-  --> $DIR/implied_bounds_in_impl.rs:35:62
-   |
-LL | fn generics_implied_multi2<T, V>() -> impl GenericTrait<T> + GenericTrait2<V> + GenericSubtrait<(), T, V>
-   |                                                              ^^^^^^^^^^^^^^^^
-
-error: this bound is implied by another bound and can be removed
-  --> $DIR/implied_bounds_in_impl.rs:45:28
-   |
-LL | fn generics_same() -> impl GenericTrait<i32> + GenericSubtrait<(), i32, ()> {}
-   |                            ^^^^^^^^^^^^^^^^^
-
-error: aborting due to 7 previous errors
-
