+++ tests/ui/implied_bounds_in_impls.rs
@@ -1,5 +1,6 @@
-#![warn(clippy::implied_bounds_in_impl)]
+#![warn(clippy::implied_bounds_in_impls)]
 #![allow(dead_code)]
+#![feature(return_position_impl_trait_in_trait)]
 
 use std::ops::{Deref, DerefMut};
 
@@ -15,7 +16,9 @@ fn deref_derefmut<T>(x: T) -> impl Deref<Target = T> + DerefMut<Target = T> {
 
 trait GenericTrait<T> {}
 trait GenericTrait2<V> {}
-// U is intentionally at a different "index" in GenericSubtrait than `T` is in GenericTrait
+// U is intentionally at a different "index" in GenericSubtrait than `T` is in GenericTrait,
+// so this can be a good test to make sure that the calculations are right (no off-by-one errors,
+// ...)
 trait GenericSubtrait<T, U, V>: GenericTrait<U> + GenericTrait2<V> {}
 
 impl GenericTrait<i32> for () {}
@@ -24,9 +27,9 @@ impl<V> GenericTrait2<V> for () {}
 impl<V> GenericSubtrait<(), i32, V> for () {}
 impl<V> GenericSubtrait<(), i64, V> for () {}
 
-fn generics_implied<T>() -> impl GenericTrait<T> + GenericSubtrait<(), T, ()>
+fn generics_implied<U, W>() -> impl GenericTrait<W> + GenericSubtrait<U, W, U>
 where
-    (): GenericSubtrait<(), T, ()>,
+    (): GenericSubtrait<U, W, U>,
 {
 }
 
@@ -43,5 +46,25 @@ fn generics_different() -> impl GenericTrait<i32> + GenericSubtrait<(), i64, ()>
 
 // i32 == i32, GenericSubtrait<_, i32, _> does imply GenericTrait<i32>, lint
 fn generics_same() -> impl GenericTrait<i32> + GenericSubtrait<(), i32, ()> {}
+
+trait SomeTrait {
+    // Check that it works in trait declarations.
+    fn f() -> impl Deref + DerefMut<Target = u8> {
+        Box::new(0)
+    }
+}
+struct SomeStruct;
+impl SomeStruct {
+    // Check that it works in inherent impl blocks.
+    fn f() -> impl DerefMut<Target = u8> {
+        Box::new(123)
+    }
+}
+impl SomeTrait for SomeStruct {
+    // Check that it works in trait impls.
+    fn f() -> impl DerefMut<Target = u8> {
+        Box::new(42)
+    }
+}
 
 fn main() {}
